+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - ClickHouse server and driver version upped to 54276.
 - Changelog.
 - Added column name to `TypeMismatchError`.
+- Query progress information.
 
 ## [0.0.5] - 2017-07-16
 ### Added